Support customizing Python module install destination
Introduce the option `VTK_MODULE_PYTHON_MODULE_DESTINATION` (effective
only when `VTK_WHEEL_BUILD` is OFF). This addition allows users to
customize the installation destination for Python modules. The purpose of
this change is to facilitate the installation of modules in a user-defined
location, particularly relevant when building VTK modules as part of a
Slicer extension.